+++ src/babel/__init__.py
@@ -46,7 +46,7 @@ from babel.schema import (
     MetaParameters,
 )
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 
 def run(